Add support for lists and nested objects
As yet, there is no exported component that allows use of nested content, other than the list field
