+++ src/components/markup/rte/utils/getSelectedNode.ts
