+++ src/index.ts
@@ -1,2 +1,6 @@
 export { Content } from './components/Content.js';
+export { Attributes } from './components/attribute/Attribute.js';
+export { String } from './components/string/String.js';
+export { Markup } from './components/markup/Markup.js';
+export { List } from './components/list/List.js';
 export { ResonanceContext } from './context/ResonanceContext.js';
